Move schema definition class into sep file
